Extract validation items and validators
